feat(web): separate layouts by route groups (app/settings/admin)
Introduce Next.js route groups to split navigation context:
- (app): Dashboard, Jobs, Job Runs, Datasets
- (settings): Connections, Uploads, Members, Billing
- (admin): Tenants, Analytics (superadmin only)

Extract shared auth helper (getAuthContext), replace DashboardHeader
with AppHeader (gear/shield icon links), add SettingsNav and AdminNav
secondary tab bars. Remove ~330 lines of duplicated auth fetch and
header boilerplate from page components. Add /members to middleware
protected routes.

Closes #100
