+++ apps/node/web/src/app/(app)/jobs/[id]/job-detail-manager.tsx
