+++ apps/node/web/src/app/(app)/jobs/jobs-manager.tsx
