--- apps/node/web/src/app/jobs/page.tsx
@@ -1,57 +1,0 @@
-import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
-
-import { DashboardHeader } from "@/app/dashboard/dashboard-header";
-import type { components } from "@/lib/api/generated";
-import { backendFetch } from "@/lib/api/server";
-import { TENANT_COOKIE, TOKEN_COOKIE } from "@/lib/auth/constants";
-import { JobsManager } from "./jobs-manager";
-
-type MeResponse = components["schemas"]["MeResponse"];
-type Job = components["schemas"]["Job"];
-
-export default async function JobsPage() {
-  const jar = await cookies();
-  const token = jar.get(TOKEN_COOKIE)?.value;
-  const tenantId = jar.get(TENANT_COOKIE)?.value;
-  if (!token || !tenantId) {
-    redirect("/signin");
-  }
-
-  const meRes = await backendFetch("/api/v1/auth/me", {
-    headers: { Authorization: `Bearer ${token}` },
-  });
-  if (!meRes.ok) {
-    redirect("/signin");
-  }
-  const me: MeResponse = await meRes.json();
-
-  let jobs: Job[] = [];
-  const jobsRes = await backendFetch("/api/v1/jobs", {
-    headers: {
-      Authorization: `Bearer ${token}`,
-      "X-Tenant-ID": tenantId,
-    },
-    cache: "no-store",
-  });
-  if (jobsRes.ok) {
-    const data: { items: Job[] } = await jobsRes.json();
-    jobs = data.items ?? [];
-  }
-
-  return (
-    <div className="min-h-screen">
-      <DashboardHeader
-        displayName={me.display_name}
-        email={me.email}
-        platformRole={me.platform_role}
-        tenants={me.tenants}
-        currentTenantId={tenantId}
-      />
-      <main className="container space-y-6 py-8">
-        <h1 className="text-2xl font-semibold tracking-tight">Jobs</h1>
-        <JobsManager initialJobs={jobs} />
-      </main>
-    </div>
-  );
-}
