+++ apps/node/web/src/components/event-summary.tsx
